fix: address remote recovery review feedback
