+++ Tests/MuxyTests/Services/Terminal/TerminalLaunchCommandTests.swift
@@ -123,37 +123,6 @@ struct TerminalLaunchCommandTests {
         #expect(script.contains("while :; do sleep 3600; done"))
     }
 
-    @Test("Remote recovery state notifies the pane")
-    @MainActor
-    func remoteRecoveryStateNotifiesPane() {
-        let recoveryToken = UUID()
-        let view = GhosttyTerminalNSView(
-            workingDirectory: "~",
-            workspaceContext: .ssh(SSHDestination(host: "prod")),
-            remoteRecoveryToken: recoveryToken
-        )
-        var events: [String] = []
-        view.onSearchEnd = { events.append("search-ended") }
-        view.onRemoteSessionRecoveryFailed = { events.append("recovery-\($0)") }
-
-        let forgedTitle = TerminalLaunchCommand.remoteReconnectRequiredTitle(recoveryToken: UUID())
-        let validTitle = TerminalLaunchCommand.remoteReconnectRequiredTitle(recoveryToken: recoveryToken)
-
-        #expect(!view.handleRemoteSessionRecoveryTitle(forgedTitle))
-        #expect(!view.processExitHandled)
-        #expect(view.handleRemoteSessionRecoveryTitle(validTitle))
-        #expect(events == ["search-ended", "recovery-true"])
-        #expect(view.isRemoteSessionRecoveryFailed)
-        #expect(view.processExitHandled)
-
-        view.retryRemoteSession()
-
-        #expect(events == ["search-ended", "recovery-true", "recovery-false"])
-        #expect(!view.isRemoteSessionRecoveryFailed)
-        #expect(!view.processExitHandled)
-        #expect(!view.handleRemoteSessionRecoveryTitle(validTitle))
-    }
-
     @Test("Remote reconnect wrapper is valid POSIX shell")
     func remoteReconnectWrapperIsValidShell() throws {
         let process = Process()
